requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx>=0.24.0
python-multipart>=0.0.6
elevenlabs>=0.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
